refactor: type StackLimit as uint and delete the ItemStack enum
The enum defined only 1/10/50/100 while the currency code already bypassed
it with (ItemStack)120u, so it bought no safety - just a cast to forget.
StackLimit is now a plain uint (default 1u) and every consumer drops its
(uint)/(int) cast.

+++ Assets/Scripts/InventorySystem/Runtime/Inventories/CharacterInventory.cs
@@ -51,7 +51,7 @@ namespace ToolSmiths.InventorySystem.Inventories
 
             void TryAddToInventory()
             {
-                var amount = Math.Min(package.Amount, (uint)package.Item.StackLimit);
+                var amount = Math.Min(package.Amount, package.Item.StackLimit);
 
                 if (StoredPackages.TryAdd(position, new Package(this, package.Item, amount)))
                     _ = package.ReduceAmount(amount);
